feat: add broadcasts to stations
