+++ client/components/index.js
@@ -14,6 +14,6 @@ export {default as Listener} from './Listener';
 export {default as MyAccount} from './MyAccount';
 export {default as NewBroadcast} from './NewBroadcast';
 export {default as SingleStation} from './SingleStation';
-export {default as Player} from './playback';
+export {default as Playback} from './Playback';
 export {default as MyStations} from './MyStations';
 export {default as NewStation} from './NewStation';
